nersion number has been changed and unused files are removed
