+++ example/App.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from "react";
 import { StatusBar } from "react-native";
-// import SocialLoginScreen from "react-native-social-login-screen";
-import SocialLoginScreen from "./lib/SocialLoginScreen";
+import SocialLoginScreen from "react-native-social-login-screen";
 
 const App = () => {
   const [isLoginButtonSpinner, setIsLoginButtonSpinner] = useState(false);
