+++ lib/components/SocialButton/SocialButton.tsx
@@ -1,6 +1,7 @@
 import * as React from "react";
-import { Text, TouchableOpacity } from "react-native";
+import { View, Text, TouchableOpacity } from "react-native";
 import Androw from "react-native-androw";
+import Spinner from "react-native-spinkit";
 /**
  * ? Local Imports
  */
@@ -14,6 +15,10 @@ interface ISocialButtonProps {
   height?: number | string;
   component?: React.ReactNode;
   loginButtonTextStyle?: any;
+  isSpinner?: boolean;
+  spinnerSize?: number;
+  spinnerType?: string;
+  spinnerColor?: string;
   onPress: () => void;
 }
 
@@ -26,19 +31,39 @@ const SocialButton = (props: ISocialButtonProps) => {
     loginButtonTextStyle,
     shadowColor = "#757575",
     backgroundColor = "#69bc4c",
+    isSpinner = false,
+    spinnerSize = 16,
+    spinnerType = "Circle",
+    spinnerColor = "#fff",
     onPress,
   } = props;
+
+  const renderSpinner = () => {
+    return (
+      <View>
+        <Spinner
+          size={spinnerSize}
+          type={spinnerType}
+          color={spinnerColor}
+          isVisible={isSpinner}
+        />
+      </View>
+    );
+  };
+
   return (
     <Androw style={_shadowStyle(shadowColor)}>
       <TouchableOpacity
         style={_container(width, height, backgroundColor)}
         onPress={onPress}
       >
-        {component || (
-          <Text style={[styles.buttonTextStyle, loginButtonTextStyle]}>
-            {text}
-          </Text>
-        )}
+        {isSpinner
+          ? renderSpinner()
+          : component || (
+              <Text style={[styles.buttonTextStyle, loginButtonTextStyle]}>
+                {text}
+              </Text>
+            )}
       </TouchableOpacity>
     </Androw>
   );
