+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-social-login-screen",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "description": "Easy Login Screen with beautiful design and fully customizable options for React Native",
   "main": "./build/dist/SocialLoginScreen.js",
   "repository": "git@github.com:WrathChaos/react-native-social-login-screen.git",
@@ -69,5 +69,8 @@
       "npx prettier --write",
       "eslint src/*.js --fix-dry-run"
     ]
+  },
+  "dependencies": {
+    "react-native-spinkit": "^1.5.1"
   }
 }
